requirements_new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blis==1.3.0
boto3==1.40.30
botocore==1.40.30
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.116.1
frozenlist==1.7.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.4.27
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.107.2
orjson==3.11.3
packaging==24.2
pathlib2==2.3.7
portalocker==3.2.0
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.1
pydantic==2.11.8
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
s3transfer==0.14.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
starlette==0.47.3
tenacity==8.5.0
thinc==8.3.6
tiktoken==0.11.0
tqdm==4.67.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
yarl==1.20.1
zstandard==0.24.0
